rename: 쿠폰 서비스 패키지 이름 수정(_제거)
